refactor modules package and rename interfaces
